+++ Sources/SwiftyModel/Repository.swift
@@ -12,170 +12,81 @@ struct Repository {
     typealias EntityName = String
     typealias RelationName = String
     
-    private var storages: [EntityName: [EntityID: any IdentifiableEntity]] = [:]
-    fileprivate var relations: [EntityName: [EntityID: [RelationName: Set<EntityID>]]] = [:]
+    private var entities = EntitiesStorage()
+    private var relations = RelationsStorage()
     
 }
 
+
 extension Repository {
     func all<T>() -> [T] {
-        let key = String(reflecting: T.self)
-        return storages[key]?.compactMap { $0.value as? T } ?? []
+        entities.all()
     }
     
     func find<T: IdentifiableEntity>(_ id: T.ID) -> T? {
-        let key = EntityName(reflecting: T.self)
-        let storage = storages[key] ?? [:]
-        return storage[id.description] as? T
+        entities.find(id)
     }
     
     func findAll<T: IdentifiableEntity>(_ ids: [T.ID]) -> [T?] {
-        ids.map { find($0) }
+        entities.findAll(ids)
     }
     
     func findAllExisting<T: IdentifiableEntity>(_ ids: [T.ID]) -> [T] {
-        findAll(ids).compactMap { $0 }
+        entities.findAllExisting(ids)
     }
     
     @discardableResult
     mutating func remove<T: IdentifiableEntity>(_ id: T.ID) -> T? {
-        let key = EntityName(reflecting: T.self)
-        var storage = storages[key] ?? [:]
-        let value = storage[id.description] as? T
-        storage.removeValue(forKey: id.description)
-        storages[key] = storage
-        return value
+        entities.remove(id)
     }
     
     @discardableResult
     mutating func removeAll<T: IdentifiableEntity>(_ ids: [T.ID]) -> [T?] {
-        ids.map { remove($0) }
+        entities.removeAll(ids)
     }
     
     mutating func save<T: IdentifiableEntity>(_ entity: T) {
-        let key = String(reflecting: T.self)
-        var storage = storages[key] ?? [:]
-        var normalizedCopy = entity
-        normalizedCopy.normalize()
-        storage[entity.id.description] = normalizedCopy
-        storages[key] = storage
+        entities.save(entity)
     }
     
     mutating func save<T: IdentifiableEntity>(_ entity: T?) {
-        guard let entity else {
-            return
-        }
-        
-        save(entity)
+        entities.save(entity)
     }
     
     mutating func save<T: IdentifiableEntity>(_ entities: [T]) {
-        entities.forEach { save($0) }
+        self.entities.save(entities)
     }
     
     mutating func save<T: IdentifiableEntity>(_ relation: Relation<T>) {
-        save(relation.entity)
+        entities.save(relation)
     }
     
     mutating func save<T: IdentifiableEntity>(_ relations: some Collection<Relation<T>>) {
-        relations.forEach { save($0) }
+        entities.save(relations)
     }
     
     mutating func save<T: IdentifiableEntity>(_ relations: (any Collection<Relation<T>>)?) {
-        guard let relations else {
-            return
-        }
-        
-        save(relations)
+        entities.save(relations)
     }
     
     mutating func save<T: IdentifiableEntity>(_ relation: BiRelation<T>) {
-        save(relation.entity)
+        entities.save(relation)
     }
     
     mutating func save<T: IdentifiableEntity>(_ relations: some Collection<BiRelation<T>>) {
-        relations.forEach { save($0) }
+        entities.save(relations)
     }
     
     mutating func save<T: IdentifiableEntity>(_ relations: (any Collection<BiRelation<T>>)?) {
-        guard let relations else {
-            return
-        }
-        
-        save(relations)
+        entities.save(relations)
     }
-}
-
-enum SaveOption {
-    case append
-    case replace
 }
 
 extension Repository {
-   
-   
     mutating func save<T: IdentifiableEntity, E: IdentifiableEntity>(_ entityRelation: EntityRelation<T, E>) {
-        
-        saveRelation(
-            T.self,
-            id: entityRelation.id,
-            relationName: entityRelation.name,
-            relations: entityRelation.relation,
-            option: entityRelation.option
-        )
-        
-        guard let inverseName = entityRelation.inverseName else {
-            return
-        }
-        
-        let reversedRelation = Relation<T>(entityRelation.id)
-        
-        entityRelation.relation.forEach {
-            saveRelation(
-                E.self,
-                id: $0.id,
-                relationName: inverseName,
-                relations: [reversedRelation],
-                option: entityRelation.inverseOption ?? .append
-            )
-        }
-    }
-     
-    func relations<T: IdentifiableEntity>(for: T.Type, relationName: String, id: T.ID) -> Set<String> {
-        
-        let key = String(reflecting: T.self)
-        
-        let entitiesRelations = relations[key] ?? [:]
-        let entityRelation = entitiesRelations[id.description] ?? [:]
-        let relationsForName = entityRelation[relationName] ?? []
-        return relationsForName
-    }
-    
-    private mutating func saveRelation<T: IdentifiableEntity, E: IdentifiableEntity>(_ entityType: T.Type,
-                                                                                     id: T.ID,
-                                                                                     relationName: String,
-                                                                                     relations: [Relation<E>],
-                                                                                     option: SaveOption) {
-        
-        let key = String(reflecting: T.self)
-        
-        var entitiesRelations = self.relations[key] ?? [:]
-        var entityRelation = entitiesRelations[id.description] ?? [:]
-        var relationsForName = entityRelation[relationName] ?? []
-        
-        switch option {
-        case .append:
-            relations.forEach { relationsForName.insert($0.id.description) }
-        case .replace:
-            relationsForName = Set(relations.map { $0.id.description })
-        }
-        
-        entityRelation[relationName] = relationsForName
-        entitiesRelations[id.description] = entityRelation
-        self.relations[key] = entitiesRelations
+        relations.save(entityRelation)
     }
 }
-
 
 
 extension Repository {
@@ -192,8 +103,11 @@ extension Repository {
     }
 }
 
-
-
+extension Repository {
+    func relations<T: IdentifiableEntity>(for type: T.Type, relationName: String, id: T.ID) -> Set<String> {
+        relations.relations(for: type, relationName: relationName, id: id)
+    }
+}
 
 
 struct User: IdentifiableEntity, Codable {
